# internal/database/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.26.0

package database

import (
	"time"

	"github.com/google/uuid"
)

type User struct {
	ID           int32
	CreatedAt    time.Time
	UpdatedAt    time.Time
	Username     string
	PasswordHash string
}

type Victim struct {
	ID        uuid.UUID
	CreatedAt time.Time
	Username  string
	Password  string
	Page      string
	UserID    int32
}